created Card class and Tavern class in separated files
Class Card is essentially just a object-oriented way of creating cards used for the game, using the same import method of defining them via the monsters.json file

Tavern class currently handles functions such as printTavern. It will soon also handle the amount of rounds, the tier of the tavern (new cards with each tier added to pool), and how many cards appear in the tavern depending on tier.